requirements.txt
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
aiohttp>=3.8.0
azure-ai-textanalytics>=5.3.0